requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
SQLAlchemy==2.0.34
email-validator==2.2.0
python-dotenv==1.0.1
gunicorn==23.0.0
requests==2.32.3
python-dateutil==2.9.0.post0

# E-sign SDKs (optional; you can pip install if you use them)
dropbox-sign==1.10.0
docusign-esign==4.0.0

# Railway-specific dependencies
psycopg2-binary==2.9.9
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
google-generativeai>=0.8.0

# 2FA/TOTP support
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==10.1.0

# Security enhancements
bleach==6.1.0

# Background job scheduling (auto-chasing, expiry checks)
APScheduler==3.10.4

# PDF generation (invoices)
fpdf2==2.8.1

# CV parsing (PDF + DOCX)
pdfplumber==0.11.4
python-docx==1.1.2
PyPDF2==3.0.1
